Extracted CovidDataChart from ChartsPage
