+++ covid-resource-guide/src/App.js
@@ -1,8 +1,5 @@
 import logo from './logo.svg';
 import './App.css';
-import ChartsInput from "./ChartsInput.jsx";
-import AjaxComponent from "./AjaxComponent.jsx";
-import { Line } from "react-chartjs-2";
 import ChartsPage from "./ChartsPage.jsx";
