+++ covid-resource-guide/src/CovidDataChart.jsx
@@ -1,0 +1,39 @@
+import { Line } from "react-chartjs-2";
+
+function CovidDataChart(){
+  var jsonString =
+    '{"ROWCOUNT":599,"COLUMNS":["fips","date","daily_cases"],"DATA":{"FIPS":["Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County","Essex County"],"DATE":["2020-03-12","2020-03-13","2020-03-14","2020-03-15","2020-03-16","2020-03-17","2020-03-18","2020-03-19","2020-03-20","2020-03-21","2020-03-22","2020-03-23","2020-03-24","2020-03-25","2020-03-26","2020-03-27","2020-03-28","2020-03-29","2020-03-30","2020-03-31","2020-04-01","2020-04-02","2020-04-03","2020-04-04","2020-04-05","2020-04-06","2020-04-07","2020-04-08","2020-04-09","2020-04-10","2020-04-11","2020-04-12","2020-04-13","2020-04-14","2020-04-15","2020-04-16","2020-04-17","2020-04-18","2020-04-19","2020-04-20","2020-04-21","2020-04-22","2020-04-23","2020-04-24","2020-04-25","2020-04-26","2020-04-27","2020-04-28","2020-04-29","2020-04-30","2020-05-01","2020-05-02","2020-05-03","2020-05-04","2020-05-05","2020-05-06","2020-05-07","2020-05-08","2020-05-09","2020-05-10","2020-05-11","2020-05-12","2020-05-13","2020-05-14","2020-05-15","2020-05-16","2020-05-17","2020-05-18","2020-05-19","2020-05-20","2020-05-21","2020-05-22","2020-05-23","2020-05-24","2020-05-25","2020-05-26","2020-05-27","2020-05-28","2020-05-29","2020-05-30","2020-05-31","2020-06-01","2020-06-02","2020-06-03","2020-06-04","2020-06-05","2020-06-06","2020-06-07","2020-06-08","2020-06-09","2020-06-10","2020-06-11","2020-06-12","2020-06-13","2020-06-14","2020-06-15","2020-06-16","2020-06-17","2020-06-18","2020-06-19","2020-06-20","2020-06-21","2020-06-22","2020-06-23","2020-06-24","2020-06-25","2020-06-26","2020-06-27","2020-06-28","2020-06-29","2020-06-30","2020-07-01","2020-07-02","2020-07-03","2020-07-04","2020-07-05","2020-07-06","2020-07-07","2020-07-08","2020-07-09","2020-07-10","2020-07-11","2020-07-12","2020-07-13","2020-07-14","2020-07-15","2020-07-16","2020-07-17","2020-07-18","2020-07-19","2020-07-20","2020-07-21","2020-07-22","2020-07-23","2020-07-24","2020-07-25","2020-07-26","2020-07-27","2020-07-28","2020-07-29","2020-07-30","2020-07-31","2020-08-01","2020-08-02","2020-08-03","2020-08-04","2020-08-05","2020-08-06","2020-08-07","2020-08-08","2020-08-09","2020-08-10","2020-08-11","2020-08-12","2020-08-13","2020-08-14","2020-08-15","2020-08-16","2020-08-17","2020-08-18","2020-08-19","2020-08-20","2020-08-21","2020-08-22","2020-08-23","2020-08-24","2020-08-25","2020-08-26","2020-08-27","2020-08-28","2020-08-29","2020-08-30","2020-08-31","2020-09-01","2020-09-02","2020-09-03","2020-09-04","2020-09-05","2020-09-06","2020-09-07","2020-09-08","2020-09-09","2020-09-10","2020-09-11","2020-09-12","2020-09-13","2020-09-14","2020-09-15","2020-09-16","2020-09-17","2020-09-18","2020-09-19","2020-09-20","2020-09-21","2020-09-22","2020-09-23","2020-09-24","2020-09-25","2020-09-26","2020-09-27","2020-09-28","2020-09-29","2020-09-30","2020-10-01","2020-10-02","2020-10-03","2020-10-04","2020-10-05","2020-10-06","2020-10-07","2020-10-08","2020-10-09","2020-10-10","2020-10-11","2020-10-12","2020-10-13","2020-10-14","2020-10-15","2020-10-16","2020-10-17","2020-10-18","2020-10-19","2020-10-20","2020-10-21","2020-10-22","2020-10-23","2020-10-24","2020-10-25","2020-10-26","2020-10-27","2020-10-28","2020-10-29","2020-10-30","2020-10-31","2020-11-01","2020-11-02","2020-11-03","2020-11-04","2020-11-05","2020-11-06","2020-11-07","2020-11-08","2020-11-09","2020-11-10","2020-11-11","2020-11-12","2020-11-13","2020-11-14","2020-11-15","2020-11-16","2020-11-17","2020-11-18","2020-11-19","2020-11-20","2020-11-21","2020-11-22","2020-11-23","2020-11-24","2020-11-25","2020-11-26","2020-11-27","2020-11-28","2020-11-29","2020-11-30","2020-12-01","2020-12-02","2020-12-03","2020-12-04","2020-12-05","2020-12-06","2020-12-07","2020-12-08","2020-12-09","2020-12-10","2020-12-11","2020-12-12","2020-12-13","2020-12-14","2020-12-15","2020-12-16","2020-12-17","2020-12-18","2020-12-19","2020-12-20","2020-12-21","2020-12-22","2020-12-23","2020-12-24","2020-12-25","2020-12-26","2020-12-27","2020-12-28","2020-12-29","2020-12-30","2020-12-31","2021-01-01","2021-01-02","2021-01-03","2021-01-04","2021-01-05","2021-01-06","2021-01-07","2021-01-08","2021-01-09","2021-01-10","2021-01-11","2021-01-12","2021-01-13","2021-01-14","2021-01-15","2021-01-16","2021-01-17","2021-01-18","2021-01-19","2021-01-20","2021-01-21","2021-01-22","2021-01-23","2021-01-24","2021-01-25","2021-01-26","2021-01-27","2021-01-28","2021-01-29","2021-01-30","2021-01-31","2021-02-01","2021-02-02","2021-02-03","2021-02-04","2021-02-05","2021-02-06","2021-02-07","2021-02-08","2021-02-09","2021-02-10","2021-02-11","2021-02-12","2021-02-13","2021-02-14","2021-02-15","2021-02-16","2021-02-17","2021-02-18","2021-02-19","2021-02-20","2021-02-21","2021-02-22","2021-02-23","2021-02-24","2021-02-25","2021-02-26","2021-02-27","2021-02-28","2021-03-01","2021-03-02","2021-03-03","2021-03-04","2021-03-05","2021-03-06","2021-03-07","2021-03-08","2021-03-09","2021-03-10","2021-03-11","2021-03-12","2021-03-13","2021-03-14","2021-03-15","2021-03-16","2021-03-17","2021-03-18","2021-03-19","2021-03-20","2021-03-21","2021-03-22","2021-03-23","2021-03-24","2021-03-25","2021-03-26","2021-03-27","2021-03-28","2021-03-29","2021-03-30","2021-03-31","2021-04-01","2021-04-02","2021-04-03","2021-04-04","2021-04-05","2021-04-06","2021-04-07","2021-04-08","2021-04-09","2021-04-10","2021-04-11","2021-04-12","2021-04-13","2021-04-14","2021-04-15","2021-04-16","2021-04-17","2021-04-18","2021-04-19","2021-04-20","2021-04-21","2021-04-22","2021-04-23","2021-04-24","2021-04-25","2021-04-26","2021-04-27","2021-04-28","2021-04-29","2021-04-30","2021-05-01","2021-05-02","2021-05-03","2021-05-04","2021-05-05","2021-05-06","2021-05-07","2021-05-08","2021-05-09","2021-05-10","2021-05-11","2021-05-12","2021-05-13","2021-05-14","2021-05-15","2021-05-16","2021-05-17","2021-05-18","2021-05-19","2021-05-20","2021-05-21","2021-05-22","2021-05-23","2021-05-24","2021-05-25","2021-05-26","2021-05-27","2021-05-28","2021-05-29","2021-05-30","2021-05-31","2021-06-01","2021-06-02","2021-06-03","2021-06-04","2021-06-05","2021-06-06","2021-06-07","2021-06-08","2021-06-09","2021-06-10","2021-06-11","2021-06-12","2021-06-13","2021-06-14","2021-06-15","2021-06-16","2021-06-17","2021-06-18","2021-06-19","2021-06-20","2021-06-21","2021-06-22","2021-06-23","2021-06-24","2021-06-25","2021-06-26","2021-06-27","2021-06-28","2021-06-29","2021-06-30","2021-07-01","2021-07-02","2021-07-03","2021-07-04","2021-07-05","2021-07-06","2021-07-07","2021-07-08","2021-07-09","2021-07-10","2021-07-11","2021-07-12","2021-07-13","2021-07-14","2021-07-15","2021-07-16","2021-07-17","2021-07-18","2021-07-19","2021-07-20","2021-07-21","2021-07-22","2021-07-23","2021-07-24","2021-07-25","2021-07-26","2021-07-27","2021-07-28","2021-07-29","2021-07-30","2021-07-31","2021-08-01","2021-08-02","2021-08-03","2021-08-04","2021-08-05","2021-08-06","2021-08-07","2021-08-08","2021-08-09","2021-08-10","2021-08-11","2021-08-12","2021-08-13","2021-08-14","2021-08-15","2021-08-16","2021-08-17","2021-08-18","2021-08-19","2021-08-20","2021-08-21","2021-08-22","2021-08-23","2021-08-24","2021-08-25","2021-08-26","2021-08-27","2021-08-28","2021-08-29","2021-08-30","2021-08-31","2021-09-01","2021-09-02","2021-09-03","2021-09-04","2021-09-05","2021-09-06","2021-09-07","2021-09-08","2021-09-09","2021-09-10","2021-09-11","2021-09-12","2021-09-13","2021-09-14","2021-09-15","2021-09-16","2021-09-17","2021-09-18","2021-09-19","2021-09-20","2021-09-21","2021-09-22","2021-09-23","2021-09-24","2021-09-25","2021-09-26","2021-09-27","2021-09-28","2021-09-29","2021-09-30","2021-10-01","2021-10-02","2021-10-03","2021-10-04","2021-10-05","2021-10-06","2021-10-07","2021-10-08","2021-10-09","2021-10-10","2021-10-11","2021-10-12","2021-10-13","2021-10-14","2021-10-15","2021-10-16","2021-10-17","2021-10-18","2021-10-19","2021-10-20","2021-10-21","2021-10-22","2021-10-23","2021-10-24","2021-10-25","2021-10-26","2021-10-27","2021-10-28","2021-10-29","2021-10-30","2021-10-31"],"DAILY_CASES":[1,5,1,4,9,12,13,18,10,34,65,101,69,39,228,217,260,141,337,336,362,355,450,517,498,411,585,520,471,511,427,403,224,578,367,505,588,229,403,425,399,259,424,299,410,343,184,143,255,237,312,254,273,100,219,111,144,161,109,159,78,56,91,75,129,79,108,460,86,166,54,108,51,77,60,53,83,112,96,48,35,66,57,59,101,107,47,11,40,34,55,28,61,41,28,11,13,33,78,34,-18,36,27,14,25,257,29,30,20,15,27,13,37,38,17,28,8,23,2,16,26,34,20,31,19,48,19,-8,-6,0,16,32,12,25,30,35,47,25,41,48,13,47,17,21,32,30,19,33,42,29,13,19,29,32,27,30,41,16,11,63,35,-3,19,36,24,19,28,19,36,41,16,25,23,44,24,15,56,25,17,23,17,29,17,15,27,24,12,33,21,26,24,30,30,21,20,25,29,27,70,30,16,21,29,36,44,59,70,35,42,46,86,69,81,71,61,92,106,111,98,118,134,146,110,142,191,185,356,162,160,246,278,254,261,147,175,145,259,357,262,293,362,256,176,689,360,421,344,656,344,207,357,388,393,341,539,337,310,385,360,325,323,154,409,317,430,294,472,580,602,516,322,437,340,432,363,505,439,382,296,567,337,372,337,418,222,303,354,345,445,287,145,254,237,389,429,507,371,309,3685,419,483,627,545,609,457,335,379,656,687,599,568,449,305,295,429,370,286,569,538,347,327,491,440,740,491,406,324,300,178,210,330,346,467,304,273,520,330,360,355,206,166,262,461,346,279,271,149,236,266,284,356,338,428,212,277,254,321,291,328,355,228,195,334,339,369,426,345,285,243,368,422,443,409,408,346,295,433,443,389,572,484,425,403,530,307,676,467,390,468,256,399,525,508,427,346,376,236,461,375,509,499,430,372,260,330,400,396,356,311,189,-2825,221,238,241,185,163,103,94,71,53,43,63,-35,-88,-34,13,51,21,5,31,-23,1,71,76,53,72,-11,-2,-16,20,57,39,35,19,18,20,15,60,22,17,32,15,21,23,9,33,11,-3,12,22,38,26,27,16,9,7,16,17,16,4,26,8,-7,12,11,-3,0,21,-17,7,1,7,11,36,30,7,41,41,51,52,39,46,45,51,28,90,74,108,58,86,51,48,84,110,119,82,66,75,84,115,116,115,95,113,108,62,156,160,166,142,164,132,113,164,151,149,233,147,71,130,131,181,146,161,103,150,46,213,130,180,142,150,111,104,27,251,161,216,198,99,67,196,146,203,158,203,98,97,169,165,150,165,155,80,32,101,141,131,59,111,83,57,84,147,134,92,97,78,35,97,125,91,30,52,13,-8,35,16,23,91,21,23,52,118,65,81,55,42,74]}}';
+  var dataObject = JSON.parse(jsonString);
+  var numRows = dataObject.ROWCOUNT;
+  var points = [];
+    for (let i = 0; i < numRows; i++) {
+      let point = {
+        x: dataObject.DATA.DATE[i],
+        y: dataObject.DATA.DAILY_CASES[i]
+      };
+      points.push(point);
+    }
+  var data = {
+      labels: dataObject.DATA.DATE,
+      datasets: [
+        {
+          label: "Daily Cases",
+          data: points,
+          backgroundColor: "rgb(255, 99, 132)",
+          borderColor: "rgba(255, 99, 132, 0.2)"
+        }
+      ]
+    };
+    var options = {
+      scales: {
+        xAxes: [
+          {
+            type: "time"
+          }
+        ]
+      }
+    };
+  return <Line data={data} options={options} />;
+
+}
+export default CovidDataChart;
